refactor: migrate from next to astro
