--- .prettierrc.json
@@ -1,8 +1,0 @@
-{
-	"printWidth": 100,
-	"useTabs": true,
-	"quoteProps": "as-needed",
-	"singleQuote": true,
-	"trailingComma": "all",
-	"endOfLine": "lf"
-}
