+++ astro.config.mjs
@@ -1,0 +1,11 @@
+import { defineConfig } from 'astro/config';
+
+import mdx from '@astrojs/mdx';
+
+import react from '@astrojs/react';
+
+// https://astro.build/config
+export default defineConfig({
+	integrations: [mdx(), react()],
+	outDir: './out',
+});
