+++ package.json
@@ -3,42 +3,41 @@
 	"version": "0.1.0",
 	"private": true,
 	"scripts": {
-		"dev": "next dev",
-		"build": "next build",
-		"start": "next start",
-		"lint": "prettier --check . && eslint_d --format=pretty src",
-		"format": "prettier --write . && eslint_d --fix --format=pretty src",
+		"dev": "astro dev",
+		"build": "astro build",
+		"lint": "prettier --check .; eslint --format=pretty src",
+		"format": "./Taskfile lint:fix",
 		"test": "mocha src/**/*.spec.ts",
 		"typecheck": "tsc --noEmit",
 		"e2e": "./Taskfile playwright:local",
 		"e2e:ui": "./Taskfile playwright:ui"
 	},
 	"dependencies": {
-		"@emotion/cache": "^11.13.1",
-		"@emotion/react": "^11.13.3",
-		"@emotion/styled": "^11.13.0",
+		"@astrojs/mdx": "^3.1.7",
+		"@astrojs/react": "^3.6.2",
 		"@radix-ui/react-accordion": "^1.2.0",
 		"@radix-ui/react-icons": "^1.3.0",
 		"@radix-ui/react-toast": "^1.2.1",
 		"@radix-ui/react-tooltip": "^1.1.2",
+		"astro": "^4.15.9",
 		"copy-to-clipboard": "^3.3.3",
+		"luxon": "^3.5.0",
 		"moment-timezone": "^0.5.45",
-		"next": "^14.2.13",
 		"prop-types": "^15.8.1",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
-		"react-toastify": "^10.0.5",
-		"recharts": "^2.12.7"
+		"react-toastify": "^10.0.5"
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.47.2",
 		"@types/chai": "^4.3.19",
 		"@types/d3-time-format": "^4.0.3",
 		"@types/ignore-styles": "^5.0.3",
+		"@types/luxon": "^3",
 		"@types/mocha": "^10.0.8",
 		"@types/node": "^20.16.5",
 		"@types/prop-types": "^15.7.13",
-		"@types/react": "^18.3.8",
+		"@types/react": "^18.3.9",
 		"@types/react-dom": "^18.3.0",
 		"chai": "^5.1.1",
 		"eslint": "^8.57.1",
@@ -46,11 +45,11 @@
 		"eslint-config-next": "14.2.13",
 		"eslint-formatter-pretty": "^6.0.1",
 		"eslint-plugin-prettier": "^5.2.1",
-		"eslint_d": "^14.0.4",
 		"ignore-styles": "^5.0.1",
 		"lodash.merge": "^4.6.2",
 		"mocha": "^10.7.3",
 		"prettier": "^3.3.3",
+		"prettier-plugin-astro": "^0.14.1",
 		"tsx": "^4.19.1",
 		"typescript": "^5.6.2"
 	},
