+++ public/guide/farmAfterSendingCBsWithTimer.png
