--- src/components/ScrollToTop/styles.module.css
@@ -1,48 +1,0 @@
-.scrollToTopButton {
-	position: fixed;
-	bottom: 20px;
-	right: 20px;
-	z-index: 100;
-	width: 50px;
-	height: 50px;
-	border-radius: 50%;
-	border: none;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	background-color: transparent;
-	cursor: pointer;
-}
-
-@keyframes zoomIn {
-	0% {
-		transform: translateX(100px) scale(0.5);
-		opacity: 0;
-	}
-
-	100% {
-		transform: scale(1);
-		opacity: 1;
-	}
-}
-
-@keyframes zoomOut {
-	0% {
-		transform: scale(1);
-		opacity: 1;
-	}
-
-	100% {
-		transform: translateX(100px) scale(0.5);
-		opacity: 0;
-	}
-}
-
-.zoomInButton {
-	animation: zoomIn 0.5s forwards;
-}
-
-.zoomOutButton {
-	animation: zoomOut 0.5s forwards;
-}
