+++ src/components/ToastMessage.tsx
@@ -1,8 +1,6 @@
-import React from 'react';
-import { ToastContainer, toast } from 'react-toastify';
+import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-// Reusable toast notification function
 export default function ToastMessage() {
 	return (
 		<div>
