--- src/pages/contract-boost-calculator/styles.module.css
@@ -1,151 +1,0 @@
-.boostBtnContainer {
-	/* width: 40vw; */
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: flex-start;
-	align-items: center;
-	gap: 1rem;
-	margin-bottom: 24px;
-}
-
-.boostBtn {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: fit-content;
-	height: fit-content;
-	/* background-color: dodgerblue; */
-	padding: 0.5rem 1rem;
-	border-radius: 10px;
-	border: none;
-	cursor: pointer;
-	color: #000;
-	transition: all 0.3s;
-	white-space: nowrap;
-}
-
-.AccordionTrigger:hover,
-.boostBtn:hover {
-	filter: brightness(90%);
-}
-
-.boostBtn:active {
-	filter: brightness(80%);
-	transform: scale(0.85);
-}
-
-.activeBoostBtn {
-	background-color: dodgerblue;
-	color: #fff;
-}
-
-.boostAndArtifactInputsContainer {
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	align-items: flex-start;
-	width: 100%;
-}
-
-.inputContainer {
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	align-items: flex-start;
-	gap: 0.5rem;
-	margin-bottom: 0.75rem;
-}
-
-.inputContainer input,
-.stonesInputsContainer select,
-.inputContainer select {
-	border-radius: 10px;
-	border: 1px solid #000;
-	padding: 0.5rem;
-}
-
-.boostAndArtifactInputsContainer div input {
-	width: 100px;
-}
-
-.stonesInputsContainer {
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	align-items: flex-start;
-	gap: 0.25rem;
-}
-
-.stonesInputsContainer select {
-	width: 60px;
-	margin: 0 16px 0 8px;
-}
-
-.main section:nth-last-child(2) {
-	/* Your styles here */
-	margin-bottom: 50px;
-}
-
-@media screen and (max-width: 768px) {
-	.stonesInputsContainer > div > div {
-		display: grid;
-		row-gap: 1rem;
-		grid-template-columns: repeat(2, auto);
-		justify-content: start;
-		align-items: center;
-	}
-}
-
-/* Accordion styles obtained from https://www.radix-ui.com/primitives/docs/components/accordion */
-
-.AccordionChevron {
-	transition: transform 300ms;
-}
-
-.AccordionTrigger {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	padding: 0.75rem;
-	border-radius: 10px;
-	border: none;
-	cursor: pointer;
-}
-
-.AccordionTrigger[data-state='open'] > .AccordionChevron {
-	transform: rotate(180deg);
-}
-
-.AccordionContent {
-	overflow: hidden;
-	border-bottom: 1px solid gray;
-	margin-bottom: 10px;
-}
-
-.AccordionContent[data-state='open'] {
-	animation: slideDown 400ms ease-out;
-}
-
-.AccordionContent[data-state='closed'] {
-	animation: slideUp 200ms ease-out;
-}
-
-@keyframes slideDown {
-	from {
-		height: 0;
-	}
-
-	to {
-		height: var(--radix-accordion-content-height);
-	}
-}
-
-@keyframes slideUp {
-	from {
-		height: var(--radix-accordion-content-height);
-	}
-
-	to {
-		height: 0;
-	}
-}
