--- src/pages/globals.css
@@ -1,49 +1,0 @@
-body {
-	max-width: 90vw;
-	margin: 0 auto;
-}
-
-main {
-	max-width: 80%;
-	margin: 2rem auto;
-}
-
-.bold {
-	font-weight: 600;
-}
-
-table {
-	border: 1px solid black;
-}
-
-table th,
-table td {
-	text-align: left;
-	vertical-align: top;
-	border: 1px solid black;
-}
-
-img {
-	object-fit: contain;
-	max-width: 75vw;
-}
-
-#guide section {
-	margin-bottom: 1rem;
-}
-
-/* .responsive-image {
-  height: auto; */
-/* This will maintain the aspect ratio */
-/* width: 100%; */
-/* This will make the image take the full width of the container */
-/* } */
-
-/* For mobile screens  */
-@media (max-width: 767px) {
-	.responsive-image {
-		width: 100%;
-		height: auto;
-		/* Or any height you want on mobile */
-	}
-}
