+++ src/pages/initiator-tools/min-fails-generator.astro
@@ -1,0 +1,8 @@
+---
+import Layout from '/layouts/Layout.astro';
+import MinFailsGenerator from '/components/initiator-tools/min-fails-generator';
+---
+
+<Layout title="Min Fails Generator">
+	<MinFailsGenerator />
+</Layout>
