+++ src/pages/initiator-tools/notin-fails-generator.astro
@@ -1,0 +1,8 @@
+---
+import Layout from '/layouts/Layout.astro';
+import NotinFailsGenerator from '/components/initiator-tools/notin-fails-generator';
+---
+
+<Layout title="Notin Fails Generator">
+	<NotinFailsGenerator />
+</Layout>
