+++ src/pages/initiator-tools/notin-message-generator.astro
@@ -1,0 +1,8 @@
+---
+import Layout from '/layouts/Layout.astro';
+import NotinMessageGenerator from '/components/initiator-tools/notin-message-generator';
+---
+
+<Layout title="Notin Message Generator">
+	<NotinMessageGenerator />
+</Layout>
