+++ tsconfig.json
@@ -1,29 +1,11 @@
 {
+	"extends": "astro/tsconfigs/strictest",
 	"compilerOptions": {
-		"rootDir": ".",
 		"baseUrl": ".",
 		"paths": {
-			"@/*": ["src/*"]
+			"/*": ["src/*"]
 		},
-		"target": "esnext",
-		"lib": ["dom", "dom.iterable", "esnext"],
-		"allowJs": true,
-		"skipLibCheck": true,
-		"strict": true,
-		"noEmit": true,
-		"esModuleInterop": true,
-		"module": "esnext",
-		"moduleResolution": "bundler",
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"jsx": "preserve",
-		"incremental": true,
-		"plugins": [
-			{
-				"name": "next"
-			}
-		]
-	},
-	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-	"exclude": ["node_modules"]
+		"jsx": "react-jsx",
+		"jsxImportSource": "react"
+	}
 }
